Provide info about changed objects from Module.
Moving states for example could affect more than just the given state.
It affects all substates and all transitions connected to the state or
any of its substates. Instead of making the App class know these details
the Model will now provide a list for every type of objects that may
need to be redrawn in the changes field.