Remove note application features
Previously applying a note title caused renaming all occurences in
the model. But this was very confusing as there was seemingly two
ways to change the trigger event of a transition and text cache
coherency was started to be a nightmare to manage. So this feature
is no more. Clicking on a text now only allows changing the assigned
notes. Caching was removed so anything in the text area is written
directly into the actual model data structure.