Prevent triggering event handlers in parent
There was a bug in the previous commit: after a state handled an event
the event processing was not stopped to evaluate guards in ancestors,
but if the parent has event handler defined for the same event it was
also triggered.

This commit fixes this behavior by disabling further signal processing
for already handled signals.