Remove signals.h dependency from everywhere
After this commit the modules in the examples folder will include a
compile time defined header file. This makes it more straightforward to
define signal classes at the application level.

The following should be passed to the C compiler:
-DSIGNAL_CLASSES_H="signal_classes.h"

signal_classes.h shall define the value for SIG_CLASS_XXXXX constants
in an enum, or in macros.